requirements.txt
aniso8601==3.0.2
asn1crypto==0.24.0
cffi==1.11.5
cryptography==2.4.2
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.1.4
django-graphql-jwt==0.2.1
django-heroku==0.3.1
django-rest-knox==3.6.0
django-storages==1.7.1
django-toolbelt==0.0.1
djangorestframework==3.9.0
drf-writable-nested==0.5.1
graphene==2.1.3
graphene-django==2.2.0
graphql-core==2.1
graphql-relay==0.4.5
gunicorn==19.9.0
idna==2.8
promise==2.2.1
psycopg2==2.7.6.1
psycopg2-binary==2.7.6.1
pycparser==2.19
PyJWT==1.7.1
pyOpenSSL==18.0.0
pytz==2017.2
Rx==1.6.1
singledispatch==3.4.0.3
six==1.12.0
static3==0.7.0
whitenoise==4.1.2
